extlinux/syslinux: fix set-kernel
Currently we can't set default kernel with extlinux/syslinux because we basically
the current implementation doesn't load the installed kernel data and when comparing
the detected ones to check with the selected kernel we'll use a data the set-kernel
command doesn't provide us. Long story short this patch brings to life the
set-kernel command for legacy bios mode bootloaders.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>